+++ DrillingRig.ConfigApp/AppControl/AinSettingsWrite/AinSettingsWritable.cs
@@ -18,8 +18,8 @@ namespace DrillingRig.ConfigApp.AppControl.AinSettingsWrite {
 		public double FlMinMin { get; set; }
 		public short IoutMax { get; set; }
 		public double FiMin { get; set; }
-		public short DacCh { get; set; }
-		public short Imcw { get; set; }
+		public ushort DacCh { get; set; }
+		public ushort Imcw { get; set; }
 		public short Ia0 { get; set; }
 		public short Ib0 { get; set; }
 		public short Ic0 { get; set; }
@@ -43,29 +43,36 @@ namespace DrillingRig.ConfigApp.AppControl.AinSettingsWrite {
 		public double KpIq { get; set; }
 		public double KiIq { get; set; }
 
-		public short AccDfDt { get; set; }
-		public short DecDfDt { get; set; }
+		public double AccDfDt { get; set; }
+		public double DecDfDt { get; set; }
 		public double Unom { get; set; }
 		public double TauFlLim { get; set; }
 		public double Rs { get; set; }
 		public double Fmin { get; set; }
-		public short TauM { get; set; }
-		public short TauF { get; set; }
-		public short TauFSet { get; set; }
-		public short TauFi { get; set; }
+
+		public double TauM { get; set; }
+		public double TauF { get; set; }
+		public double TauFSet { get; set; }
+		public double TauFi { get; set; }
+
 		public short IdSetMin { get; set; }
 		public short IdSetMax { get; set; }
+
 		public BytesPair UchMin { get; set; }
 		public BytesPair UchMax { get; set; }
+
 		public BytesPair Reserved50 { get; set; }
 		public BytesPair Reserved51 { get; set; }
+
 		public int Np { get; set; }
 		public int NimpFloorCode { get; set; }
 		public AinTelemetryFanWorkmode FanMode { get; set; }
+
 		public double UmodThr { get; set; }
-		public short EmdecDfdt { get; set; }
+		public double EmdecDfdt { get; set; }
 		public short TextMax { get; set; }
 		public short ToHl { get; set; }
+
 		public bool Ain1LinkFault { get; set; }
 		public bool Ain2LinkFault { get; set; }
 		public bool Ain3LinkFault { get; set; }
